Add activities nearby points search

+++ src/main/kotlin/ua/com/radiokot/flybys/analysis/model/ActivityFlyby.kt
@@ -3,7 +3,18 @@ package ua.com.radiokot.flybys.analysis.model
 import ua.com.radiokot.flybys.strava.activities.model.Activity
 
 data class ActivityFlyby(
+        /**
+         * Activity that is Flyby to the source one
+         */
         val activity: Activity,
+
+        /**
+         * Indices of [activity] points that are nearby to source activity
+         */
         val nearbyPointIndices: List<Int>,
-        val correlationPercent: Double
+
+        /**
+         * Percent of source and [activity] location-time correlation
+         */
+        val correlationPercent: Int
 )
